Fix multitouch
