+++ Sources/UIApplication+handleSDLEvents.swift
@@ -16,8 +16,8 @@ extension UIApplication {
         while SDL_PollEvent(&e) == 1 {
 
             #if os(Android)
-            if let uievent = UIEvent.from(e) {
-                sendEvent(uievent)
+            if let uiEvent = UIEvent.from(e) {
+                sendEvent(uiEvent)
                 break
             }
             #endif
@@ -150,127 +150,3 @@ extension SDL_Event {
         return TimeInterval(self.common.timestamp) / 1000
     }
 }
-
-extension UIEvent {
-    static func from(_ event: SDL_Event) -> UIEvent? {
-        switch SDL_EventType(event.type) {
-        case SDL_FINGERDOWN:
-            let newTouch = UITouch(
-                touchId: Int(event.tfinger.fingerId),
-                at: CGPoint(
-                    x: CGFloat(event.tfinger.x),
-                    y: CGFloat(event.tfinger.y)
-                ),
-                timestamp: event.timestampInSeconds
-            )
-
-            if
-                let firstExistingEvent = UIEvent.activeEvents.first,
-                let _ = firstExistingEvent.allTouches?.first(where: {$0.touchId == event.tfinger.fingerId})
-            {
-                // Found a matching event, adding current touch to it and returning
-                firstExistingEvent.allTouches?.insert(newTouch)
-                return firstExistingEvent
-            } else {
-                // No matching event found, creating a new one
-                return UIEvent(touch: newTouch)
-            }
-
-        case SDL_FINGERMOTION:
-            if
-                let firstExistingEvent = UIEvent.activeEvents.first,
-                let matchingTouch = firstExistingEvent.allTouches?.first(where: {$0.touchId == event.tfinger.fingerId})
-            {
-
-                matchingTouch.timestamp = event.timestampInSeconds
-                matchingTouch.phase = .moved
-                matchingTouch.updateAbsoluteLocation(.from(event.tfinger))
-                return firstExistingEvent
-            } else {
-                return nil
-            }
-
-        case SDL_FINGERUP:
-            if
-                let firstExistingEvent = UIEvent.activeEvents.first,
-                let matchingTouch = firstExistingEvent.allTouches?.first(where: {$0.touchId == event.tfinger.fingerId})
-            {
-                matchingTouch.timestamp = event.timestampInSeconds
-                matchingTouch.phase = .ended
-                return firstExistingEvent
-            } else {
-                return nil
-            }
-
-        default:
-            return nil
-        }
-    }
-}
-
-extension UIEvent: CustomStringConvertible {
-    public var description: String {
-        let text = "Event with touches: "
-        guard let allTouches = self.allTouches else {
-            return text + "none"
-        }
-        let descriptions = allTouches.map { $0.description }
-        return text + descriptions.joined(separator: ",\n")
-    }
-}
-
-extension UITouch: CustomStringConvertible {
-    public var description: String {
-        return "TouchID: \(self.touchId), timestamp: \(self.timestamp)"
-    }
-}
-
-#if os(Android)
-import JNI
-
-@_cdecl("Java_org_libsdl_app_SDLActivity_onNativeTouchUIKit")
-public func onNativeTouch(
-    env: UnsafeMutablePointer<JNIEnv>,
-    view: JavaObject,
-    touchDeviceID: JavaInt,
-    pointerFingerID: JavaInt,
-    action: JavaInt,
-    x: JavaFloat,
-    y: JavaFloat,
-    pressure: JavaFloat,
-    timestamp: JavaLong
-) {
-    guard let eventType = SDL_EventType.eventFrom(androidAction: action)
-    else { return }
-
-    var event = SDL_Event(tfinger:
-        SDL_TouchFingerEvent(
-            type: eventType.rawValue,
-            timestamp: UInt32(timestamp), // ensure this is in ms
-            touchId: Int64(touchDeviceID), // I think this is the "Touch Device ID" which should always be 0, but check this
-            fingerId: Int64(pointerFingerID),
-            x: x / Float(UIScreen.main.scale),
-            y: y / Float(UIScreen.main.scale),
-            dx: 0,
-            dy: 0,
-            pressure: pressure
-        )
-    )
-
-    // add the event to SDL's event stack
-    // don't use SDL_PushEvent because it overrides `event.timestamp` with its own:
-    SDL_PeepEvents(&event, 1, SDL_ADDEVENT, 0, 0)
-}
-
-extension SDL_EventType {
-    public static func eventFrom(androidAction: JavaInt) -> SDL_EventType? {
-        switch androidAction {
-        case 0, 5: return SDL_FINGERDOWN
-        case 1, 6: return SDL_FINGERUP
-        case 2: return SDL_FINGERMOTION
-        default: return nil
-        }
-    }
-}
-
-#endif
